Add web UI integration of plan-day synthesizer (Phase 8, #22)
Surface the plan-day synthesizer in the web app as a "Suggest a Day"
panel: the user clicks, sees the archetype plans, and loads one into the
existing drag-and-drop editor for fine-tuning.

- Port the synthesizer's business logic to a pure lib/synthesizer.ts
  module (synthesizeDay: inputs -> archetype plans), shared by the CLI and
  the web app. cli/commands/plan-day.ts is now a thin IO wrapper that
  gathers inputs (week store, inbox, catalog, weather, workout), calls the
  shared synthesizer, and handles --apply + human formatting. Types are
  re-exported for backward compatibility.
- Add app/api/plan-day route (Node runtime) that gathers server-side
  inputs and runs the shared synthesizer around the canvas's current busy
  events and plan blocks.
- Add SuggestDayModal: fetches plans, renders the archetype cards
  side-by-side (rationale + block list + scheduled/free stats), and loads
  the chosen plan's blocks into the editor. Wired into the header.
